[Bleeding] Add new CreateReason for End Portal creation. Adds BUKKIT-4145
Previously, there was no specific reason to indicate that the portal being
created was an End Portal instead of a Nether Portal or a custom portal. This
adds a new CreateReason, PLACED_ENDER_EYE, to indicate that the portal was
created when the last Eye of Ender was placed into an End Portal frame.