[Bleeding] Add API for damage modifiers in EntityDamageEvent. Adds BUKKIT-347, BUKKIT-4104
This commit adds API for the enchantment, armor, potion and other modifications to damage done to an entity.  These damage reductions are each modifiable via a getter and a setter.  This addition allows for more accurate modification and monitoring of damage done to/by an entity, as it displays the final damage done as well.